Alteração Enviar E-mail
Implementado função para enviar e-mail quando um idGit não existir na tabela PROJECTS, na base ENG

+++ ProjectMetricAPI/EngProjectMetricsIntegration.cs
@@ -1,8 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace ProjectMetricAPI
 {
@@ -13,7 +9,7 @@ namespace ProjectMetricAPI
         public DateTime executed_at { get; set; }
         public double coverage { get; set; }
         public int total { get; set; }
-        public int percentege { get; set; }
+        public double percentege { get; set; }
 
     }
 }
